+++ src/browser.rs
@@ -3,5 +3,6 @@ pub mod fetch;
 pub mod service;
 pub mod url;
 pub mod util;
+pub mod web_storage;
 
 pub use url::{Url, DUMMY_BASE_URL};