+++ src/browser/service.rs
@@ -1,4 +1,3 @@
 pub mod fetch;
 pub mod routing;
 pub mod storage;
-pub mod web_storage;