+++ src/browser/service/storage.rs
@@ -10,7 +10,7 @@ extern crate serde_json;
 
 pub type Storage = web_sys::Storage;
 
-#[deprecated(since = "0.7.0", note = "Use web_storage::get_local_storage")]
+#[deprecated(since = "0.7.0", note = "Use `LocalStorage` or `SessionStorage`.")]
 #[allow(clippy::module_name_repetitions)]
 pub fn get_storage() -> Option<Storage> {
     web_sys::window()
@@ -21,7 +21,10 @@ pub fn get_storage() -> Option<Storage> {
 }
 
 /// Create a new store, from a serializable data structure.
-#[deprecated(since = "0.7.0", note = "use WebStorage.store_data")]
+#[deprecated(
+    since = "0.7.0",
+    note = "Use `LocalStorage::insert` or `SessionStorage::insert`."
+)]
 pub fn store_data<T>(storage: &Storage, name: &str, data: &T)
 where
     T: serde::Serialize,
@@ -33,7 +36,10 @@ where
 }
 
 /// Load a store, to a deserializable data structure.
-#[deprecated(since = "0.7.0", note = "use WebStorage.load_data")]
+#[deprecated(
+    since = "0.7.0",
+    note = "Use `LocalStorage::get` or `SessionStorage::get`."
+)]
 pub fn load_data<T>(storage: &Storage, name: &str) -> Option<T>
 where
     T: serde::de::DeserializeOwned,